Set up 32-bit build under Linux.

+++ Benchmarks/bm_EM.cpp
@@ -12,7 +12,7 @@ static void em_mousie(benchmark::State& state)
 	std::uniform_real_distribution<double> u01(0, 1);
 
 	const unsigned int num_dimensions = 2;
-	const int64_t sample_size = state.range(0);
+	const unsigned int sample_size = static_cast<unsigned int>(state.range(0));
 	const double face_radius = 1;
 	const double ear_radius = 0.3;
 	const unsigned int num_components = 3; // face, left ear, right ear
@@ -24,7 +24,7 @@ static void em_mousie(benchmark::State& state)
 
 	Eigen::MatrixXd data(num_dimensions, sample_size);
 	std::vector<unsigned int> classes(sample_size);
-	for (int64_t i = 0; i < sample_size; ++i) {
+	for (unsigned int i = 0; i < sample_size; ++i) {
 		const unsigned int k = component_distr(rng);
 		classes[i] = k;
 		const double phi = 2 * PI * u01(rng);
